+++ ouroboros-network-api/src/Ouroboros/Network/PeerSelection/PeerSharing.hs
@@ -4,7 +4,7 @@
 
 module Ouroboros.Network.PeerSelection.PeerSharing
   ( PeerSharing (..)
-  , combinePeerInformation
+  , combinePeerSharing
   , encodePortNumber
   , decodePortNumber
   , encodeRemoteAddress
@@ -18,8 +18,6 @@ import           Data.Aeson.Types (FromJSON (..), ToJSON (..), Value (..),
 import qualified Data.Text as Text
 import           GHC.Generics (Generic)
 import           Network.Socket (PortNumber, SockAddr (..))
-import           Ouroboros.Network.PeerSelection.PeerAdvertise
-                     (PeerAdvertise (..))
 import           Text.Read (readMaybe)
 
 -- | Is a peer willing to participate in Peer Sharing? If yes are others allowed
@@ -29,11 +27,9 @@ import           Text.Read (readMaybe)
 --
 -- NOTE: This information is only useful if P2P flag is enabled.
 --
-data PeerSharing = NoPeerSharing -- ^ Peer does not participate in Peer Sharing
-                                 -- at all
-                 | PeerSharingPrivate -- ^ Peer participates in Peer Sharing but
-                                      -- its address should be private
-                 | PeerSharingPublic -- ^ Peer participates in Peer Sharing
+data PeerSharing = PeerSharingDisabled -- ^ Peer does not participate in Peer Sharing
+                                       -- at all
+                 | PeerSharingEnabled -- ^ Peer participates in Peer Sharing
   deriving  (Eq, Show, Read, Generic)
 
 instance FromJSON PeerSharing where
@@ -46,21 +42,13 @@ instance FromJSON PeerSharing where
 instance ToJSON PeerSharing where
   toJSON = String . Text.pack . show
 
--- Combine a 'PeerSharing' value and a 'PeerAdvertise' value into a
--- resulting 'PeerSharing' that can be used to decide if we should
--- share or not the given Peer. According to the following rules:
+-- | Combine two 'PeerSharing' values
 --
--- - If no PeerSharing value is known then there's nothing we can assess
--- - If a peer is not participating in Peer Sharing ignore all other information
--- - If a peer said it wasn't okay to share its address, respect that no matter what.
--- - If a peer was privately configured with DoNotAdvertisePeer respect that no matter
--- what.
---
-combinePeerInformation :: PeerSharing -> PeerAdvertise -> PeerSharing
-combinePeerInformation NoPeerSharing      _                  = NoPeerSharing
-combinePeerInformation PeerSharingPrivate _                  = PeerSharingPrivate
-combinePeerInformation PeerSharingPublic  DoNotAdvertisePeer = PeerSharingPrivate
-combinePeerInformation _                         _           = PeerSharingPublic
+-- 'PeerSharingDisabled' is the absorbing element
+combinePeerSharing :: PeerSharing -> PeerSharing -> PeerSharing
+combinePeerSharing PeerSharingDisabled _ = PeerSharingDisabled
+combinePeerSharing _ PeerSharingDisabled = PeerSharingDisabled
+combinePeerSharing _ _                   = PeerSharingEnabled
 
 encodePortNumber :: PortNumber -> CBOR.Encoding
 encodePortNumber = CBOR.encodeWord16 . fromIntegral