+++ ouroboros-network-framework/sim-tests/Test/Ouroboros/Network/Server2/Sim.hs
@@ -153,7 +153,6 @@ tests =
 -- Server tests
 --
 
-
 prop_unidirectional_Sim :: ClientAndServerData Int
                         -> Property
 prop_unidirectional_Sim clientAndServerData =