+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/Diffusion/Node.hs
@@ -366,23 +366,25 @@ run blockGeneratorArgs limits ni na tracersExtra tracerBlockFetch =
       where
         encodeData _ NtNVersionData { ntnDiffusionMode, ntnPeerSharing } =
           case ntnDiffusionMode of
-            InitiatorOnlyDiffusionMode         -> case ntnPeerSharing of
-              NoPeerSharing      -> CBOR.TList [CBOR.TBool False, CBOR.TInt 0]
-              PeerSharingPrivate -> CBOR.TList [CBOR.TBool False, CBOR.TInt 1]
-              PeerSharingPublic  -> CBOR.TList [CBOR.TBool False, CBOR.TInt 2]
-            InitiatorAndResponderDiffusionMode -> case ntnPeerSharing of
-              NoPeerSharing      -> CBOR.TList [CBOR.TBool True, CBOR.TInt 0]
-              PeerSharingPrivate -> CBOR.TList [CBOR.TBool True, CBOR.TInt 1]
-              PeerSharingPublic  -> CBOR.TList [CBOR.TBool True, CBOR.TInt 2]
-        decodeData _ bytes = case bytes of
-          CBOR.TList [CBOR.TBool False, CBOR.TInt 0] -> Right (NtNVersionData InitiatorOnlyDiffusionMode NoPeerSharing)
-          CBOR.TList [CBOR.TBool False, CBOR.TInt 1] -> Right (NtNVersionData InitiatorOnlyDiffusionMode PeerSharingPrivate)
-          CBOR.TList [CBOR.TBool False, CBOR.TInt 2] -> Right (NtNVersionData InitiatorOnlyDiffusionMode PeerSharingPublic)
+            InitiatorOnlyDiffusionMode         ->
+              let peerSharing = case ntnPeerSharing of
+                    PeerSharingDisabled -> 0
+                    PeerSharingEnabled  -> 1
+               in CBOR.TList [CBOR.TBool False, CBOR.TInt peerSharing]
+            InitiatorAndResponderDiffusionMode ->
+              let peerSharing = case ntnPeerSharing of
+                    PeerSharingDisabled -> 0
+                    PeerSharingEnabled  -> 1
+               in CBOR.TList [CBOR.TBool True, CBOR.TInt peerSharing]
 
-          CBOR.TList [CBOR.TBool True, CBOR.TInt 0] -> Right (NtNVersionData InitiatorAndResponderDiffusionMode NoPeerSharing)
-          CBOR.TList [CBOR.TBool True, CBOR.TInt 1] -> Right (NtNVersionData InitiatorAndResponderDiffusionMode PeerSharingPrivate)
-          CBOR.TList [CBOR.TBool True, CBOR.TInt 2] -> Right (NtNVersionData InitiatorAndResponderDiffusionMode PeerSharingPublic)
-          _                -> Left (Text.pack "unversionedDataCodec: unexpected term")
+        toPeerSharing :: Int -> PeerSharing
+        toPeerSharing 0 = PeerSharingDisabled
+        toPeerSharing 1 = PeerSharingEnabled
+        toPeerSharing _ = error "toPeerSharing: out of bounds"
+
+        decodeData _ (CBOR.TList [CBOR.TBool False, CBOR.TInt a]) = Right (NtNVersionData InitiatorOnlyDiffusionMode (toPeerSharing a))
+        decodeData _ (CBOR.TList [CBOR.TBool True, CBOR.TInt a])  = Right (NtNVersionData InitiatorAndResponderDiffusionMode (toPeerSharing a))
+        decodeData _ _                                            = Left (Text.pack "unversionedDataCodec: unexpected term")
 
     args :: Diff.Arguments (NtNFD m) NtNAddr (NtCFD m) NtCAddr
     args = Diff.Arguments