+++ ouroboros-network-api/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Breaking changes
 
+- Remote `PeerSharingPrivate` option from the `PeerSharing` data type.
+- Rename `NoPeerSharing` and `PeerSharingPublic` to `PeerSharingDisabled` and
+  `PeerSharingEnabled`, respectively.
+- Add new `NodeToNodeV_13` that encodes and decodes the updated `PeerSharing` flag data
+  type.
+
 ### Non-breaking changes
 
 * Restructured `decodeTerm` to prevent an impossible case and eliminate the